Using `ReferenceCountingFunctionsPolicy' to types with no free func.
Some commits ago, taningia started to manage memory of almost all its
types with reference counting, this commit makes our binding use it.

A little bug was found in pybindgen that was reported with a patch
attached and this patch must be applied to make it work properly.

launchpad bug: https://bugs.launchpad.net/pybindgen/+bug/568192